[develop]指标页面地图定位

+++ app/src/main/java/lxing14/software/edu/nju/cn/waterqualitymonitoring/constant/CommonConstant.java
@@ -12,10 +12,10 @@ public class CommonConstant {
     public static final String STN_ID = "stnId";
     public static final String LATITUDE = "latitude";
     public static final String LONGITUDE = "longitude";
-    public static final double LATITUDE_OF_BJ = 39.904989;
-    public static final double LONGITUDE_OF_BJ = 116.405285;
-    public static final double LATITUDE_OF_NJ = 31.864978;
-    public static final double LONGITUDE_OF_NJ = 118.87056;
+    public static final float LATITUDE_OF_BJ = 39.904989f;
+    public static final float LONGITUDE_OF_BJ = 116.405285f;
+    public static final float LATITUDE_OF_NJ = 31.864978f;
+    public static final float LONGITUDE_OF_NJ = 118.87056f;
     public static final int DEFAULT_DAY = 15;
 
 }
